特殊スキル計算を関数に切り出す
