+++ src/components/skillArea/SkillArea.tsx
@@ -1,20 +1,28 @@
-import React, { useContext } from 'react';
+import { useContext } from 'react';
 
 import { Grid } from '@mui/material';
 import Button from '@mui/material/Button';
 
-import { StatusInputFields, CharacterStatus } from 'interface/Status';
 import {
   CharacterStatusContext,
   InputStatusContext,
 } from 'contexts/StatusContext';
-import { STATUS } from '../../constants/constants';
+import { STATUS } from 'constants/constants';
 import {
   getInputStatus,
   calculateDisplayStatus,
   resetBradScraperStatus,
   resetSpecialSkillStatus,
-} from '../../util/StatusUtil';
+} from 'util/StatusUtil';
+import {
+  calcFairyKingBlessingPhy,
+  calcFairyKingBlessingMag,
+  calcArchangelBlessing,
+  calcBlessedAzureAegis,
+  calcEvilGodCurse,
+  calcWisdomKingInspirationYang,
+  calcWisdomKingProtectionYin,
+} from './specialSkills';
 
 const SkillArea = () => {
   const characterContext = useContext(CharacterStatusContext);
@@ -28,7 +36,7 @@ const SkillArea = () => {
   }
   // 必要な関数や状態を抽出するための分割代入
   const { characterStatus, updateCharacter } = characterContext;
-  const { inputStatus, updateInputStatus } = inputContext;
+  const { inputStatus } = inputContext;
 
   // ブラッドスクレイパーのスキルを計算する
   const handleBradScraper = () => {
@@ -48,29 +56,9 @@ const SkillArea = () => {
     getInputStatus(characterStatus, inputStatus, updateCharacter);
     calculateDisplayStatus(characterStatus, updateCharacter);
 
-    let newCharacterStatus = { ...characterStatus };
+    // 物理の妖精王の祝福のスキルを計算
+    calcFairyKingBlessingPhy(characterStatus, updateCharacter);
 
-    newCharacterStatus[STATUS.POW].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.POW].totalWithoutItem +
-        newCharacterStatus[STATUS.POW].card +
-        newCharacterStatus[STATUS.POW].allVita +
-        newCharacterStatus[STATUS.POW].vita +
-        newCharacterStatus[STATUS.POW].scroll +
-        newCharacterStatus[STATUS.POW].canSeal +
-        newCharacterStatus[STATUS.POW].bradScraper) *
-        0.3,
-    );
-    newCharacterStatus[STATUS.ATK].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.ATK].totalWithoutItem -
-        newCharacterStatus[STATUS.POW].totalWithoutItem +
-        newCharacterStatus[STATUS.ATK].scroll +
-        (newCharacterStatus[STATUS.POW].displayStatus -
-          newCharacterStatus[STATUS.POW].totalWithoutItem +
-          newCharacterStatus[STATUS.POW].specialSkill) *
-          2) *
-        0.3,
-    );
-    updateCharacter(newCharacterStatus);
     // 表示用ステータスを計算
     calculateDisplayStatus(characterStatus, updateCharacter);
   };
@@ -83,28 +71,9 @@ const SkillArea = () => {
     getInputStatus(characterStatus, inputStatus, updateCharacter);
     calculateDisplayStatus(characterStatus, updateCharacter);
 
-    let newCharacterStatus = { ...characterStatus };
+    // 魔法の妖精王の祝福のスキルを計算
+    calcFairyKingBlessingMag(characterStatus, updateCharacter);
 
-    newCharacterStatus[STATUS.INT].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.INT].totalWithoutItem +
-        newCharacterStatus[STATUS.INT].card +
-        newCharacterStatus[STATUS.INT].allVita +
-        newCharacterStatus[STATUS.INT].vita +
-        newCharacterStatus[STATUS.INT].scroll +
-        newCharacterStatus[STATUS.INT].canSeal) *
-        0.3,
-    );
-    newCharacterStatus[STATUS.MAT].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.MAT].totalWithoutItem +
-        newCharacterStatus[STATUS.MAT].scroll +
-        (newCharacterStatus[STATUS.INT].displayStatus -
-          newCharacterStatus[STATUS.INT].totalWithoutItem +
-          newCharacterStatus[STATUS.INT].specialSkill) *
-          2) *
-        0.3,
-    );
-
-    updateCharacter(newCharacterStatus);
     // 表示用ステータスを計算
     calculateDisplayStatus(characterStatus, updateCharacter);
   };
@@ -117,113 +86,9 @@ const SkillArea = () => {
     getInputStatus(characterStatus, inputStatus, updateCharacter);
     calculateDisplayStatus(characterStatus, updateCharacter);
 
-    let newCharacterStatus = { ...characterStatus };
+    // 大天使の加護のスキルを計算
+    calcArchangelBlessing(characterStatus, updateCharacter);
 
-    // HP・SP
-    newCharacterStatus[STATUS.HP].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.HP].totalWithoutItem +
-        newCharacterStatus[STATUS.HP].scroll) *
-        4,
-    );
-    newCharacterStatus[STATUS.SP].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.SP].totalWithoutItem +
-        newCharacterStatus[STATUS.SP].scroll) *
-        4,
-    );
-
-    // POW・INT・SPD・VIT・LUK
-    newCharacterStatus[STATUS.POW].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.POW].totalWithoutItem +
-        newCharacterStatus[STATUS.POW].card +
-        newCharacterStatus[STATUS.POW].allVita +
-        newCharacterStatus[STATUS.POW].vita +
-        newCharacterStatus[STATUS.POW].scroll +
-        newCharacterStatus[STATUS.POW].canSeal +
-        newCharacterStatus[STATUS.POW].bradScraper) *
-        0.2,
-    );
-    newCharacterStatus[STATUS.INT].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.INT].totalWithoutItem +
-        newCharacterStatus[STATUS.INT].card +
-        newCharacterStatus[STATUS.INT].allVita +
-        newCharacterStatus[STATUS.INT].vita +
-        newCharacterStatus[STATUS.INT].scroll +
-        newCharacterStatus[STATUS.INT].canSeal) *
-        0.2,
-    );
-    newCharacterStatus[STATUS.SPD].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.SPD].totalWithoutItem +
-        newCharacterStatus[STATUS.SPD].card +
-        newCharacterStatus[STATUS.SPD].allVita +
-        newCharacterStatus[STATUS.SPD].vita +
-        newCharacterStatus[STATUS.SPD].scroll +
-        newCharacterStatus[STATUS.SPD].canSeal) *
-        0.2,
-    );
-    newCharacterStatus[STATUS.VIT].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.VIT].totalWithoutItem +
-        newCharacterStatus[STATUS.VIT].card +
-        newCharacterStatus[STATUS.VIT].allVita +
-        newCharacterStatus[STATUS.VIT].vita +
-        newCharacterStatus[STATUS.VIT].scroll +
-        newCharacterStatus[STATUS.VIT].canSeal) *
-        0.2,
-    );
-    newCharacterStatus[STATUS.LUK].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.LUK].totalWithoutItem +
-        newCharacterStatus[STATUS.LUK].card +
-        newCharacterStatus[STATUS.LUK].allVita +
-        newCharacterStatus[STATUS.LUK].vita +
-        newCharacterStatus[STATUS.LUK].scroll +
-        newCharacterStatus[STATUS.LUK].canSeal) *
-        0.2,
-    );
-
-    // ATK・DEF・MAT・MDF
-    newCharacterStatus[STATUS.ATK].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.ATK].totalWithoutItem -
-        newCharacterStatus[STATUS.POW].totalWithoutItem +
-        newCharacterStatus[STATUS.ATK].scroll +
-        (newCharacterStatus[STATUS.POW].displayStatus -
-          newCharacterStatus[STATUS.POW].totalWithoutItem +
-          newCharacterStatus[STATUS.POW].specialSkill) *
-          2) *
-        0.2,
-    );
-    newCharacterStatus[STATUS.MAT].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.MAT].totalWithoutItem +
-        newCharacterStatus[STATUS.MAT].scroll +
-        (newCharacterStatus[STATUS.INT].displayStatus -
-          newCharacterStatus[STATUS.INT].totalWithoutItem +
-          newCharacterStatus[STATUS.INT].specialSkill) *
-          2) *
-        0.2,
-    );
-    newCharacterStatus[STATUS.DEF].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.DEF].totalWithoutItem +
-        newCharacterStatus[STATUS.DEF].scroll +
-        (newCharacterStatus[STATUS.VIT].displayStatus -
-          newCharacterStatus[STATUS.VIT].totalWithoutItem +
-          newCharacterStatus[STATUS.VIT].specialSkill) *
-          2) *
-        0.2,
-    );
-    const selectInt =
-      newCharacterStatus[STATUS.INT].displayStatus +
-      newCharacterStatus[STATUS.INT].specialSkill;
-    const selectVit =
-      newCharacterStatus[STATUS.VIT].displayStatus +
-      newCharacterStatus[STATUS.VIT].specialSkill;
-    const selectStatus = selectInt > selectVit ? selectInt : selectVit;
-    newCharacterStatus[STATUS.MDF].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.MDF].totalWithoutItem -
-        newCharacterStatus[STATUS.INT].totalWithoutItem * 15 +
-        newCharacterStatus[STATUS.MDF].scroll +
-        selectStatus * 2) *
-        0.2,
-    );
-
-    updateCharacter(newCharacterStatus);
     // 表示用ステータスを計算
     calculateDisplayStatus(characterStatus, updateCharacter);
   };
@@ -236,18 +101,9 @@ const SkillArea = () => {
     getInputStatus(characterStatus, inputStatus, updateCharacter);
     calculateDisplayStatus(characterStatus, updateCharacter);
 
-    let newCharacterStatus = { ...characterStatus };
+    // 祝福の蒼盾のスキルを計算
+    calcBlessedAzureAegis(characterStatus, updateCharacter);
 
-    newCharacterStatus[STATUS.VIT].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.VIT].totalWithoutItem +
-        newCharacterStatus[STATUS.VIT].card +
-        newCharacterStatus[STATUS.VIT].allVita +
-        newCharacterStatus[STATUS.VIT].vita +
-        newCharacterStatus[STATUS.VIT].scroll +
-        newCharacterStatus[STATUS.VIT].canSeal) *
-        0.3,
-    );
-    updateCharacter(newCharacterStatus);
     // 表示用ステータスを計算
     calculateDisplayStatus(characterStatus, updateCharacter);
   };
@@ -260,32 +116,9 @@ const SkillArea = () => {
     getInputStatus(characterStatus, inputStatus, updateCharacter);
     calculateDisplayStatus(characterStatus, updateCharacter);
 
-    let newCharacterStatus = { ...characterStatus };
+    // 明王の鼓舞・陽のスキルを計算
+    calcWisdomKingInspirationYang(characterStatus, updateCharacter);
 
-    // HP・SP
-    newCharacterStatus[STATUS.HP].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.HP].totalWithoutItem +
-        newCharacterStatus[STATUS.HP].scroll) *
-        2,
-    );
-    newCharacterStatus[STATUS.SP].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.SP].totalWithoutItem +
-        newCharacterStatus[STATUS.SP].scroll) *
-        2,
-    );
-
-    //SPD
-    newCharacterStatus[STATUS.SPD].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.SPD].totalWithoutItem +
-        newCharacterStatus[STATUS.SPD].card +
-        newCharacterStatus[STATUS.SPD].allVita +
-        newCharacterStatus[STATUS.SPD].vita +
-        newCharacterStatus[STATUS.SPD].scroll +
-        newCharacterStatus[STATUS.SPD].canSeal) *
-        0.3,
-    );
-
-    updateCharacter(newCharacterStatus);
     // 表示用ステータスを計算
     calculateDisplayStatus(characterStatus, updateCharacter);
   };
@@ -298,32 +131,9 @@ const SkillArea = () => {
     getInputStatus(characterStatus, inputStatus, updateCharacter);
     calculateDisplayStatus(characterStatus, updateCharacter);
 
-    let newCharacterStatus = { ...characterStatus };
+    // 明王の守護・陰のスキルを計算
+    calcWisdomKingProtectionYin(characterStatus, updateCharacter);
 
-    // HP・SP
-    newCharacterStatus[STATUS.HP].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.HP].totalWithoutItem +
-        newCharacterStatus[STATUS.HP].scroll) *
-        2,
-    );
-    newCharacterStatus[STATUS.SP].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.SP].totalWithoutItem +
-        newCharacterStatus[STATUS.SP].scroll) *
-        2,
-    );
-
-    //SPD
-    newCharacterStatus[STATUS.SPD].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.SPD].totalWithoutItem +
-        newCharacterStatus[STATUS.SPD].card +
-        newCharacterStatus[STATUS.SPD].allVita +
-        newCharacterStatus[STATUS.SPD].vita +
-        newCharacterStatus[STATUS.SPD].scroll +
-        newCharacterStatus[STATUS.SPD].canSeal) *
-        0.3,
-    );
-
-    updateCharacter(newCharacterStatus);
     // 表示用ステータスを計算
     calculateDisplayStatus(characterStatus, updateCharacter);
   };
@@ -336,19 +146,9 @@ const SkillArea = () => {
     getInputStatus(characterStatus, inputStatus, updateCharacter);
     calculateDisplayStatus(characterStatus, updateCharacter);
 
-    let newCharacterStatus = { ...characterStatus };
+    // 邪神の呪詛のスキルを計算
+    calcEvilGodCurse(characterStatus, updateCharacter);
 
-    newCharacterStatus[STATUS.LUK].specialSkill = Math.floor(
-      (newCharacterStatus[STATUS.LUK].totalWithoutItem +
-        newCharacterStatus[STATUS.LUK].card +
-        newCharacterStatus[STATUS.LUK].allVita +
-        newCharacterStatus[STATUS.LUK].vita +
-        newCharacterStatus[STATUS.LUK].scroll +
-        newCharacterStatus[STATUS.LUK].canSeal) *
-        0.3,
-    );
-
-    updateCharacter(newCharacterStatus);
     // 表示用ステータスを計算
     calculateDisplayStatus(characterStatus, updateCharacter);
   };
